opencv-ocr/requirements.txt
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
python-dotenv==1.0.1
